feat: Remove handler.go and move into main.go
